+++ tests/FSharpx.Tests/MonoidTests.fs
@@ -3,10 +3,10 @@
 open NUnit.Framework
 open FsCheck
 open FsCheck.NUnit
-open FSharpx.Functional
+open FSharpx
 open FSharpx.Collections
 open FSharpx.Tests.Properties
-open FSharpx.Functional.Monoid
+open FSharpx.Monoid
 
 type ByteStringGen =
     static member ByteStringArb =