+++ tests/FSharpx.Tests/OptionTests.fs
@@ -2,8 +2,8 @@
 
 open System
 open NUnit.Framework
-open FSharpx.Functional
-open FSharpx.Functional.Option
+open FSharpx
+open FSharpx.Option
 
 [<Test>]
 let ``kleisli composition``() =