+++ tests/FSharpx.Tests/ReaderTest.fs
@@ -2,7 +2,7 @@
 
 open System
 open System.Threading
-open FSharpx.Functional.Reader
+open FSharpx.Reader
 open NUnit.Framework
 open FsUnit