+++ tests/FSharpx.Tests/ZipListTests.fs
@@ -3,7 +3,6 @@
 open FsUnit
 open NUnit.Framework
 open FSharpx
-open FSharpx.Functional
 
 [<Test>]
 let zipping() =